+++ src/host_test.h
@@ -22,12 +22,12 @@ static inline unsigned long millis(void)
 }
 
 extern uint8_t get_inputs(void);
-extern void blit_image(uint8_t x,
-		       uint8_t y,
+extern void blit_image(int16_t x,
+		       int16_t y,
 		       const uint8_t *img,
 		       uint8_t flags);
-extern void blit_image_frame(uint8_t x,
-			     uint8_t y,
+extern void blit_image_frame(int16_t x,
+			     int16_t y,
 			     const uint8_t *img,
 			     uint8_t nr,
 			     uint8_t flags);
